Moved socketio and apirouter out of index.js into their own files.
